fixup! fixup! test(@ledgerhq/coin-stellar): add test for getAccountShape
